・ページを完全に削除：ページを完全に削除後、ページにはってあったコンテンツの権限つきショートカットが削除されていないため、削除するように修正。 ・ページを完全に削除：「ブロックとして配置していない該当ルームのコンテンツを親ルームがあれば、そちらへ移動、なければ完全に削除。」しなければならないのに、親ルームがあっても完全に削除していたため修正。

+++ app/Model/Block.php
@@ -538,7 +538,7 @@ class Block extends AppModel
 		$block_id = $block['Block']['id'];
 		$content_id = $block['Block']['content_id'];
 		if(isset($block['Content'])) {
-			$content = $block;
+			$content['Content'] = $block['Content'];
 		} else {
 			$content = $Content->findById($content_id);
 		}
@@ -589,7 +589,7 @@ class Block extends AppModel
 						}
 					}
 
-					if($content['Content']['is_master'] && $all_delete == _ON &&
+					if(isset($content['Content']) && $content['Content']['is_master'] && $all_delete == _ON &&
 						$page['Page']['room_id'] == $content['Content']['room_id'] && method_exists($class_name, 'delete')) {
 						// 削除アクション
 						$ret = $class->delete($content);
@@ -598,17 +598,18 @@ class Block extends AppModel
 						}
 					}
 				}
+				if(isset($content['Content'])) {
+					if(!$content['Content']['is_master'] ||
+						($all_delete == _ON && $page['Page']['room_id'] == $content['Content']['room_id'])) {
 
-				if(!$content['Content']['is_master'] ||
-					($all_delete == _ON && $page['Page']['room_id'] == $content['Content']['room_id'])) {
-
-					// 権限が付与されたショートカットか、ショートカットではない場合
-					// コンテンツがなくてもエラーとしない(コンテンツ一覧からコンテンツを削除後にブロック削除を行うとエラーとなるため)
-					$Content->delete($block['Block']['content_id']);
-				} else if(isset($parent_room_id) && $all_delete == NC_DELETE_MOVE_PARENT && $content['Content']['room_id'] != $parent_room_id) {
-					// 親のルームの持ち物に変換し、親のルーム内の権限を付与したショートカットを解除
-					if(!$Content->cancelShortcutParentRoom($content_id, $parent_room_id)) {
-						return false;
+						// 権限が付与されたショートカットか、ショートカットではない場合
+						// コンテンツがなくてもエラーとしない(コンテンツ一覧からコンテンツを削除後にブロック削除を行うとエラーとなるため)
+						$Content->delete($block['Block']['content_id']);
+					} else if(isset($parent_room_id) && $all_delete == NC_DELETE_MOVE_PARENT && $content['Content']['room_id'] != $parent_room_id) {
+						// 親のルームの持ち物に変換し、親のルーム内の権限を付与したショートカットを解除
+						if(!$Content->cancelShortcutParentRoom($content_id, $parent_room_id)) {
+							return false;
+						}
 					}
 				}
 			}
